implement fixed income bindings for DayCount and ZeroCouponBond
